Move ImageSerializedField to ov_wag/serializers.pyp
