+++ authors/models.py
@@ -7,17 +7,7 @@ from wagtail.snippets.edit_handlers import SnippetChooserPanel
 from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.snippets.models import register_snippet
 from modelcluster.fields import ParentalKey
-from rest_framework.fields import Field
-
-
-class ImageSerializedField(Field):
-    def to_representation(self, value):
-        return {
-            'url': value.file.url,
-            'title': value.title,
-            'width': value.width,
-            'height': value.height,
-        }
+from ov_wag.serializers import ImageSerializedField
 
 
 class AuthorsOrderable(Orderable):
